implement support for scheduling local non-compile jobs.
For example: you can add a symlink named "meinproc4" pointing
to icecc to your path, and it will automatically schedule
the job and then invoke meinproc4 on it locally. 

the advantage is that you can use cmake -j50 without having
to worry that there are going to be 50 meinproc jobs killing
your machine.

same could be done for moc, btw.
CCMAIL: kretz@kde.org

svn path=/branches/work/icecream-make-it-cool/; revision=689106