Hack for compiling with clang even on older icecream remotes.
Older icecream versions had /usr/bin/{gcc|g++} hardcoded when
invoking the compiler from the environment on the remote. In order
to build on such remotes, if the environment also contains clang,
include a wrapper around /usr/bin/{gcc|g++} that will either
invoke real gcc or clang, depending on an argument added by icecream.

Not entirely elegant, but it works, and the usual gcc-only case should
be unaffected.


svn path=/trunk/icecream/; revision=1304629