avoid -Wcast-align with clang
As I understand the gcc manpage, it warns with -Wcast-align only if the cast
would actually be a problem on the particular platform, so in other words it
never warns on the usual platforms. Clang on the other hand warns always,
and there are in fact a number of places where such casts do take place.
If these are actual problems in practice, I'll leave fixing them to somebody
who actually runs and compiles icecream on such a platform.