install clang/clang++ wrappers symlinks conditionally again
The idea to make them unconditional is flawed, as it the clang->icerun symlink
can be easily confused by the real compiler executable (e.g. in configure scripts
checking for the specific compiler), but it doesn't work like one if the real
compiler is not installed. Make them again optional, and add a configure option
to force them.
This reverts commit 517330435ff963d35ea1481dbba8b985aaadfe09 and adds the option.