Enable icecc to find compilers outside PATH
Teach find_compiler and friends to conditionally use the original
compiler name. This enables users to have (cross-)compilers outside
PATH, and is especially useful in combination with ccache and
CCACHE_PREFIX. In this case no directory of symbolic links is needed
nor does the (cross-)compiler have to be in the PATH.

Accomplish this by saving away the original, full compiler path/name
and use it in cases when the original name looks like an absolute or
relative path.

Signed-off-by: David Vest <davve@opera.com>