Revert "Add docbook2xman to the possible binary names in configure"
TL;DR: I was to hasty with my previous change.

The archlinux binary is called docbook2man, not docbook2_x_man. I tested my
change with the last release where it is sadly not yet fixed, and the build was
broken in there.

Then I went to git to create a patch and I made what I did, and then the build
succeeded. I was misled because it was not actually fixed by change, but by
another provided by Eike Ziller as follows:

    commit a40bae096bd51f328d6ff299077c5530729b0580
    Author: Eike Ziller <github@eikeziller.de>
    Date:   Tue Jun 18 22:55:36 2013 +0200

        Mac/brew has docbook2man instead of docbook-to-man

Hence my change was not really needed, and it currently does not seem to serve
any purpose unless a distribution pops up with such a name, but we all hope for
the better...

This reverts commit 1f2b07abb4ca63a680eafebb7ea0b2fe361dada1.