do not disable a node just because it can't chroot, if it's actually local
This is possibly a leftover from old code. If a node cannot chroot, it is
still perfectly usable for local builds. Hardcoding maxJobs to 0 seems
to serve no purpose, the info about chroot availability is known, and
it is checked in both is_elligible and pick_server(). It can cause
slow remote host to be selected over fast not-fully-loaded local host,
and moreover it can lead to a long timeout if no usable remote host
is found and the code in empty_queue() simply delays the jobs (until
a suitable remote host is found).