avoid changing to uid/gid that are already in effect
Switching to chroot or installing environment require either getuid()==0
or cap_ng support, but creating environment can run even under normal
user, and while setuid() and setgid() for the active user are no-op,
setgroups(0,NULL) is possibly not and could fail (and thus make
creating the environment fail). Since the right uid/gid are already
in effect, assume that in this effect the other groups are correct
as well.