Fix creation of native environment for gcc
In 96552a293a55d15d7179285a70868f0a55c9a6ea, the default compiler for
MacOS was changed and because of that the code could no longer assume
that gcc is the true unless clang is explicitly specified. For that
reason the commit explicitly set 'is_clang' to false when 'gcc' is
requested as compiler.

However it did so inside an if expression causing the expression to
evaluate to false always and without incrementing the extrafiles
pointer. This in turn caused the rest of create_native to believe gcc
is an extra file to be added to the tarball. No such file normally
exists so the creation of a native environment failed.

Fix this issue by moving the is_clang assignments outside the if
expressions, making the code slightly longer but clearer.