Don't take first scheduler that responds, wait for full timeout
When there is a scheduler on the local machine, that server will always
respond "instantly".  Other schedulers on the network will can not
respond as fast due to network lag.  As a result the code previously
would always select the scheduler on the local machine even if the
selection algorithm should select a different one.