Fix building with precompiled headers.
This commit tweaks the handling of precompiled headers to fix building
Qt 5.5 (other versions probably also affected).

The command line is something like:

    $(CXX) -c -include .pch/Qt5WebKitWidgets <more stuff>

On disk, the precompiled headers are in a directory called
`.pch/Qt5WebKitWidgets.gch`.  The logic in icecream is a bit too
strict as it checks for the existence of both paths:

    .pch/Qt5WebKitWidgets
    .pch/Qt5WebKitWidgets.gch

The change here is to check for either of these directories. The
reason for the check for the path without "gch" in it is for
handling header files.

Signed-off-by: tombailo <tom_bailey@btinternet.com>