Fix crash in Daemon::answer_client_requests
iceccd was regularly crashing on my PC, and investigation showed the
cause to be using an invalid iterator in Daemon::answer_client_requests.
This function calls create_env_finished for each entry in
native_environments, but create_env_finish will sometimes erase the
entry, leaving the calling function with an invalid iterator.

This commit fixes the crash by moving the erase() up to
answer_client_requests where it can be done safely