Fix broken pipe errors caused by a race condition in the daemon.
The client assumes that all file data will be sent to the daemon before
the daemon responds with a CompileResult msg.  If the daemon spawns the
compiler and the compiler waits to read all of the file data from stdin
before terminating, then the daemon sends the correct protocol
messages.  However, if the compiler dies before reading all of the file
data from stdin, then the daemon assumes the compiler crashed.
That was causing two problems.

First, the daemon was closing the client socket before the client had
finished writing all of the file data.  That was causing the client to
receive a "Broken pipe" error when calling flush_writebuf().  The client
was checking to see if an error message was sent from the daemon instead
of checking for a CompileResult.  However, the daemon was sending
neither.  This made it hard to troubleshoot the problem on the client
side.

Second, the daemon was returning an exit status equivalent to "compiler
crashed" when in reality the compiler ran fine but terminated before the
daemon expected it to.  This meant the daemon was never sending a
CompileResult even though one was available.

There were two fixes.  The first fix was to make the daemon wait for all
file input to be received from the client before it checks to see if the
compiler terminated.  Waiting for all file input makes sure the client
is in a state where it is ready to receive a CompileResult.  The second
fix was to make the daemon detect when the compiler terminates before
file input is received and treat that as a normal termination, instead
of a crash, so that stdout and stderr can be sent back to the client.