travis: Add a build description.
Support a basic build, install, and test invocation for
the travis-ci.org continuous integration service.

Also build on macOS. The systems are fairly different in
terms of system services and kernel interface, so it's worth
checking both platforms.

This commit only runs tests on Linux, since I haven't figured
out how to run them in the travis macOS environment.

Running the tests requires sudo privelages to set the chroot
capabilities, so we're forced to use the slower sudo-capable
linux vm image instead of a container.