do not run binaries from the build directory when running tests
This reverts the logic from ceeff64cfb and b654aa1690. The tests are meant
to run from the installation instead from the build, in order to check
the installed setup as closely as possible. Moreover, the second commit
here basically reverted the first by adding more complexity on top, instead
of reverting it as such.