using gcc's -fdirectives-only the same way we use clang's -frewrite-includes
It's almost the same and again it solves the problem of remote compilations
reporting warnings in macro expansions that are not reported without
proprocesed sources. Interestingly, this problem currently in practice
doesn't exist anyway because of the stderr workaround which forces local
rebuild. Still, avoiding the local recompile is worth it.