move the travis macosx build before all checks
I don't know how separate the handling of different platforms is, but
the macosx build takes quite a long time, and if they are attempted to be
queued in in the order given, then it's better for the macosx build to be
sooner. Also, the ordering logic now may be "normal builds first, test builds
afterwards".