break an infinite loop when trying to connect to a just-gone scheduler
After getting the scheduler's broadcast, it'd be remembered in schedname,
and the incorrect condition in try_get_scheduler() would keep it looping
and trying to select a scheduler, eventually selecting this no longer
available scheduler. And attempting to connect to the scheduler would fail,
but would also extend the timeout, thus try_get_scheduler() would try again,
and so on.
Introduced by 47ba035d4d.