buildargv: early return if argument NULL (codacy)
This removes one scope to address a couple of codacy warnings:


The scope of the variable 'arg' can be reduced.
153

char *arg;

The scope of the variable 'squote' can be reduced.
155

int squote = 0;

The scope of the variable 'dquote' can be reduced.
156

int dquote = 0;

The scope of the variable 'bsquote' can be reduced.
157

int bsquote = 0;

The scope of the variable 'argc' can be reduced.
158

int argc = 0;

The scope of the variable 'maxargc' can be reduced.
159

int maxargc = 0;