expandargv: fix check for dynamic allocation of argument vector
When the code interpolates the contents of response files, the
argument vector is reallocated to the new size. This only works
if it was dynamically allocated once before -- we do not want to
mess with the argv memory given to us by the init code.

The code tried to detect this with a flag, but that was never
written to, leading to than one more dynamic allocations -- one
for each response file.