Add support for clang plugin aliases
Check the optional environment variable ICECC_PLUGIN_ALIAS for
a comma separated list of aliases that translates from local
plugin names to names in the tarball. If an alias was found for
a used plugin the remote host will load the plugin from the aliased
file name instead of the local file name provided on the command
line.

Example:
export ICECC_PLUGIN_ALIAS='../plugins/libFoo.so=/usr/lib/libBar.so'
icecc clang -Xclang -load -Xclang ../plugins/libFoo.so -c t.cc

The remote parameters for the plugin will then be translated to:
-Xclang -load -Xclang /usr/lib/libBar.so