scheduler: distribute probe jobs
There are cases where a compile node is stuck on a really low
speed rating, when it was slow in the past. It might have recovered
by now, but it's not getting any new jobs to update the speed rating.

This patch implements a simple way to solve this by bypassing the
usual best server selection on up to 5% of all jobs and instead
assigns a server that hasn't got a job in a long time.