change the default for the envbasedir from /tmp/ to /var/tmp
According to issue#184 cleanup tools may clean up /tmp, which may remove
an installed environment without iceccd noticing it. As this is inherently
a race condition I don't think there's a clean recovery for this, so just say
that icecream's files are its own and nobody is supposed to mess with them.