improve heuristics deciding whether to build locally or distribute
Clients now connect to the daemon as early as possible, and daemons
send the number of clients to the scheduler. So if this is going
to be a large parallel build, scheduler should know fairly early
and prefer distributing. If it's just a configure run or a single
file changed, the number will stay low and scheduler will slightly
prefer local builds. There's unfortunately no reliable way to detect
whether a single client is not just the very first client of a large build.

The protocol changes are meant to be as a part of protocol 39.