Load calculation and job distribution improvements (#412)
* load: calculate idleLoad directly from idle ticks

No need to infer it from the other load values.

* load: make memory load a function of total system memory

The current way of calculating the memory load just clamps the load
to 0 for any modern system with a reasonable (more than 128MB) amount
of free memory.

* load: simplify load calculation

Simply report the max between CPU and memory load to the scheduler.

* speed: throttle job assignment

Newer CPUs include stuff like dynmic frequency ramping (aka TurboMode) and
SMT, which breaks the assumption that all slots are equally fast. Add a
small correction factor to account for that.

* scheduler: distribute probe jobs

There are cases where a compile node is stuck on a really low
speed rating, when it was slow in the past. It might have recovered
by now, but it's not getting any new jobs to update the speed rating.

This patch implements a simple way to solve this by bypassing the
usual best server selection on up to 5% of all jobs and instead
assigns a server that hasn't got a job in a long time.