Add support for Zstandard compression of the content
Zstandard (https://zstd.net) is a modern compression library also of the
Lempel-Ziv family of compressors. Compared to LZO, if given the same CPU
time, zstd will compress better. Testing with the "lzop" and "zstd"
command-line tools indicates that the same-CPU-time level compression
for a 7241 kB preprocessed source lies between zstd levels 2 and 3 (LZO
takes 916 µs; zstd level 2 takes 753 µs and level 3 takes 1024 µs).
Level 2 compresses this file to 1028 kB, while level 3 reaches 929 kB,
compared to LZO's 1581 kB.

However, when testing a full compilation of a large-ish source code (not
including linking), we find that CPU time is the bottleneck, shared
between the compressor and the preprocessors. The results of a best-of-5
run on this source are:

	lzo:    22.989s
	zstd1:  20.421s
	zstd2:  20.554s
	zstd3:  20.630s
	zstd4:  20.762s
	zstd5:  21.355s
	zstd6:  21.341s

Therefore, this commit chooses level 1 as the default. The 11% gain in
build time is caused the reduction in both CPU contention and the
network transmission -- zstd level 1 compresses the above source to 1029
kB, which is almost 35% smaller than the LZO content.