Stop forcing the use of -pipe in the compiler
The -pipe flag does not work when passing additional assembly
files with -Wa. GCC is not smart enough to add an additional '-'
argument to ensure the assembler still reads from stdin when there are
additional files passed through -Wa.

Additionally, as a matter of policy, I think it's wrong for icecc
to enforce using -pipe and should be left to the individual build
systems to make this decision.

Besides that, in this day and age, the performance benefits of
-pipe are dubious at best.