Add support for additional assembly files passed through gcc
The Linux kernel has started passing an extra assembly file to all
compilation steps in order to add globally available inline macros.
It does this by passing an extra file using -Wa. Currently this causes
icecc to force the build locally which means compiling the kernel
cannot be distributed.

In order to support this in icecc, we add any such files to
the extra files in the environment.