Additional Assembler Files for Linux Kernel (#430)
* Add unit tests for assembler options

This is to ensure we don't break existing functionality
in subsequent patches.

* Clean up assembly argument handling to process all arguments

This patch should have no functional changes besides handling
a few more esoteric cases of assembler arguments.

The -Wa arguments are split by comma and analyzed one by one with
a new helper function.

* Check for assembler options that have two arguments

According to the gas help file there are two options that
have two arguments where the second argument does not start
with a hyphen (--debug-prefix-map OLD=NEW and --defsym SYM=VAL).
With the current code, these arguments will be forced to run
locally, even though there's no good reason for that.

* Ensure assembler argument files run locally

Any assembler option beginning with @ specifies more options that
must be in a file. Instead of trying to parse all of those arguments,
we just force that file to run locally.

* Stop forcing the use of -pipe in the compiler

The -pipe flag does not work when passing additional assembly
files with -Wa. GCC is not smart enough to add an additional '-'
argument to ensure the assembler still reads from stdin when there are
additional files passed through -Wa.

Additionally, as a matter of policy, I think it's wrong for icecc
to enforce using -pipe and should be left to the individual build
systems to make this decision.

Besides that, in this day and age, the performance benefits of
-pipe are dubious at best.

* Add support for additional assembly files passed through gcc

The Linux kernel has started passing an extra assembly file to all
compilation steps in order to add globally available inline macros.
It does this by passing an extra file using -Wa. Currently this causes
icecc to force the build locally which means compiling the kernel
cannot be distributed.

In order to support this in icecc, we add any such files to
the extra files in the environment.