Various fixes related to Clear Linux (#431)
* Also include base libraries for AVX2 and AVX512 builds

glibc supports searching for libraries in subdirs according to the CPU
features. On x86-64, those are "haswell" and "avx512_1", as in:

$ ldd /usr/bin/icecc | grep haswell
        libstdc++.so.6 => /usr/lib64/haswell/avx512_1/libstdc++.so.6
        libc.so.6 => /usr/lib64/haswell/libc.so.6
        libm.so.6 => /usr/lib64/haswell/avx512_1/libm.so.6

With this change, we'll include the base library or libraries in the
environment. In the system above, we'll include /usr/lib64/libm.so.6,
/usr/lib64/haswell/libm.so.6 in addition to the libm listed in ldd's
output.

* Fix finding ld.so.conf on Clear Linux

Clear Linux has a "stateless" /etc, meaning that all files there are
supposed to belong to the user. The ld.so.cache file is instead in
/var/cache/ldconfig.

* Fix warning printed when /lib or /lib64 are symlinks

icecc-create-env was printing:
[...]
adding file /usr/bin/objcopy
adding file /etc/ld.so.conf=/tmp/icecc_ld_so_confTsSNqs
cp: -r not specified; omitting directory '/lib'
cp: -r not specified; omitting directory '/lib64'
creating d7801039695fbb794e026bdaf287276c.tar.gz

The -p option means to preserve ownership, timestamps, etc., but doesn't
tell cp to operate on the symlink itself. The -P option does.