force local rebuild if remote is missing file
There are cases where a remote compilation is failing due to a missing
file, that we can't detect with limited amount of work before kicking
off the compile. One such occasion is the Linux kernel including a file
in a inline ASM statement in a C file via the incbin directive.

If we detect a remote error with a missing file, force a local recompile.
This should only be hit in a very limited number of cases, so there is no
need for a more clever workaround.