do not require absolute path in debug tests with absolute path
It may be because of clang's -fdebug-compilation-dir, but gdb here
now shows the path only as debug/debug2.cpp and not the full path.
As long as the actual gdb output works (and thus gdb can find the file),
this doesn't actually matter, so do not require the full path.