try local recompile when local cpp invocation fails
When fdirectives-only is used the local cpp can fail on some macro constructs
that are incompatible with this mode of preprocessing. Currently this fails
the build, as it's not a remote error, which would get recovered by a local
rebuild, but the local side failing.

Instead of failing the build, trigger a local rebuild. The failure rate is
so small that any more sophisticated recovery seems to be wasted effort. This
leads to an additional local recompile in case of broken code, that fails in
the cpp, but I don't think that matters, as performance of failing builds is
probably not as important.

The tests need to account for the local fallback being triggered by a
failing cpp invocation.