improve cross-compile handling for clang in icecc-create-env
One clang binary can be used to compile code for different architectures.
One way to select the target architecture is the name of the compiler:
A symlink with the target triple, e.g. armv7-unknown-linux-gnueabihf-clang.

Add support for this in icecc-create-env by calling the clang binary as
specified on the command-line when querying other tools and headers.
This way the correct tools and headers for the desired architecture are
added to the environment.