remove special PCH handling with -include and -include-pch
As far as I can tell, this is mostly just needless trouble, and the purpose
of the code is unclear and hard to understand. AFAICT the extension check
means that it's mostly skipped anyway (because most headers have
the extension starting with 'h'). And there are several commits
in the history that try to fix this somehow only breaking something else.

Without handling these options in any special way this should work in all
cases, except for the case when the -include(-pch) option is required
(i.e. the source does not include the headers from which the PCH has been
built, so the PCH is the only source of that code). Since we do not
ship PCHs to remote nodes and assume the code will build fine even without
the PCH, this will obviously break. But although that mode of operation is
supported by GCC, I consider it to be weird. And if this actually breaks
things for somebody who uses it, they can complain so that it actually
becomes clear what the problem is and how it can possibly be handled.