keep order of remote arguments, especially -gsplit-dwarf (#435)
This is based on pull requests #425. Both gcc and clang
have their peculiarities when handling -gsplit-dwarf depending
on the exact order of this and other -g arguments. So keep the order
the same and do not handle any of those arguments specially.

This technically changes the protocol, but let's keep it as part
of the recent version 41 change.