handle EWOULDBLOCK, especially in flush_writebuf() (#456,#408)
Man pages for send and read say that EAGAIN and EWOULBLOCK are basically
the same and that they should be checked both. Also, make sure
log_perror() actually refers to the proper errno and not possibly
to one from select(). This hopefully avoids the error messages
from flush_writebuf() when the network gets overloaded.