try to strip binaries in icecc-create-env
It used to be the case that it did 'strip -s' on the binaries, but
f7562625a4a9 reverted it because in #154 somebody said that doing
so corrupted the binaries there, for whatever reason.
As now icecream automatically packages compiler in any location,
this now again matters a bit more, as now it's easy to use icecream
with a custom compiler build, so the "niche" case is now more likely
(and I use it). And in that case this makes a big difference
in the tarball size.
Instead of using 'strip -s', which strips "all", use only --strip-unneeded
or -g (= --strip-debug), which should be safe.