increase flush_writebuf() timeout to 30s and log timeout (#456,#408)
I've finally managed to reliably reproduce the problems with flush_writebuf()
writing error messages and forcing local compilation (building LibreOffice's
Clang compilerplugins with -g). It turns out that it really is the case
that the sending times out. This means that in those cases the network is
so overloaded that it doesn't manage to send files in 20 seconds.

I don't think there is really that much to do about it. I've increased
the timeout slightly to 30s, but if just the network transfer takes that
long, then presumably a local recompile is faster anyway. So also add
a log message saying it's a timeout and I guess that's it.