make icecc-create-env faster with Clang built as shared libs
With GCC or statically built Clang this doesn't matter, but if the compiler
consists of a number of libs and those generally depend on each other,
then add_files() gets called recursively, with many duplicates. It
actually checks for duplicates, using is_contained(), but apparently
doing path manipulations manually in bash is not very fast. This
optimization saves about 2/3 of the time.