make receiving environment more robust
The daemon was actually handling pretty poorly the cases when it couldn't
handle properly the environment it had received, e.g. when the tarball
was compressed in a way it couldn't decompress. Where "poorly" means
it easily crashed. Also handling of who deletes the Msg object was
unconsistent and error handling of handle_file_chunk_env() was broken
as well.