prevent getting stuck on "No suitable host found, delaying"
Sometimes the scheduler simply can't find any suitable host, e.g.
when asking for a remote for an environment that none of the available
nodes support, or maybe there's no node to have the required minimal
version or required feature. In that case the scheduler would needlessly
delay until it'd time out. So just bail our immediatelly in case
there's no node that could possibly handle the job.