detect better when unpacking of environment actually was successful
There wasn't any detection of whether the unpacking child process
finished successfully or not. Moreover one Jenkins build managed
to run into a state when the child finished before the daemon,
which resulted in EPIPE in the daemon when trying to write more data
(tarball had extra data at the end?).