limit number of local preprocessor invocations up to local cpus available (#285)
Apparently with large enough parallel builds the local preprocessor calls
can cause a high local load (I can't reproduce, but then I don't use that
large builds and possibly clang's -frewrite-includes is faster than
gcc's preprocessor). So extend the local file locking to automatically
use slots up to the number of local cpus(=threads) available. Using
local locking is preferred to using JobLocalBeginMsg to the daemon
because it's an internal detail and so it shouldn't e.g. show in icemon,
it should(?) also be a bit faster and simpler.
As a side effect this also changes runs without the local daemon available
to serialize to more than one job.