Allow restricting network access to a specific interface
By default, both the daemon and the scheduler listen on all network
interfaces, with no restrictions regarding the origin of a connection
request. However, there are cases in which we may want to allow access
from a certain network interface only. For example, we may want to put
the icecream cluster into its own VPN.

This commit adds an option for both the daemon and the scheduler to
listen only on a specific interface. This is implemented in the
following way: if the user specifies an interface name on the command
line, the IP address of that interface is looked up and the socket is
bound to that address.

Caveats:
* No more than one interface can be specified.
* If an interface has more than one address, only the first one is
considered.
* The use of IPv4 is hardcoded (but this is the case all over the
icecream code base).