icecc-create-env: ensure that target directories of symlinks get created
Fixes #464.

On some distributions such as Fedora, /lib64 is a symbolic link to
/usr/lib64. When a file is added that purportedly resides under /lib64,
the subsequent copying step fails because the (dangling) lib64 symlink
exists in the destination tree but its target usr/lib64 does not.

This commit ensures that all relevant directories are created in the
destination tree.

It seems that using `file` and `cut` is the only POSIX-conforming way of
determining the target of a symbolic link from a shell script -
`readlink` is GNU-specific.