create_native(): remove superfluous lstat() call
This call is not needed. On the one hand, execv() will test for the
file's existence and accessibility anyway; on the other hand, the
information retrieved by lstat() might become outdated before execv() is
called ("time of check to time of use", TOCTTOU).

So we just remove the call to lstat().