remove icecc-create-env's handling of 'weird multilib setups' (#464)
This comes from 2008's ab934f2236, which says it's for gentoo, but
doesn't say more. It doesn't seem actually needed AFAICT, and it breaks
Fedora with /lib64 being a symlink, and it's unclear what this was
trying to solve or whether that is still a problem. So remove
and if this change breaks it for something, at least they'll provide
information on why this would be needed.