always make daemons listen on the loopback interface
If --interface is used to restrict on what interface to listen, explicitly
listen also on loopback. This generally shouldn't be needed, because there
is the unix socket, but if the build runs in a chroot (e.g. with 'osc build'
from the openSUSE build service), then the unix socket is not reachable
and the client falls back to using the loopback.
The scheduler shouldn't need any such mechanism, daemons shouldn't use
loopback when talking to the scheduler.