handle flags for Clang modules
Clang modules are handled the same way like precompiled headers.
The basic assumption is that the same code should in practice compile
equally with or without the module, and so for remote builds
we simply strip the options and compile normally, which will save
the hassle of having to transfer the modules. As I understand it
Clang compiles modules in the background on demand as needed,
so remote compilations not having anything to do with modules
should be completely harmless.