remove environment to be used for compilation if it's broken
This handles the case when the unpacked environments are in a location
that may get cleaned up by something else than iceccd (e.g. /var/tmp
locations may get cleaned up by the system's regular cleanup of old
files). If iceccd detects that the environment is not usable, it
should be removed and the scheduler notified, otherwise this led
to repeated failed compilation attempts on that node.