use TCP_USER_TIMEOUT if available
We use SO_KEEPALIVE to detect network failures, but that works only
if the connection is idle. If we send a request to e.g. a scheduler
while the network is down, the timeout can be in practice 15 minutes.
And TCP_USER_TIMEOUT controls that timeout, so use something more
sensible.