prioritize jobs with higher priority (lower nice value)
Until now, the scheduler didn't care about the nice value
of the compile job, which meant that running a low-priority build
would take slots even if a normal build was started.
Now the scheduler takes this into account, so it's possible
to run 'nice make' somewhere in the background and running elsewhere
'make' will get served as soon as possible.