downgrade message about no scheduler to a warning
While not having a scheduler should not be the usual case,
it's still perfectly fine not to have a scheduler and have just
the local daemon deal with a local build. This function gets
called unconditionally e.g. when a job finishes, and so the log
will get spammed by these messages even though not having
a scheduler is not an error.